% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{call.size}
\alias{call.size}
\title{Find the number of functions and variables (counting multiplies) in a call}
\usage{
call.size(call)
}
\arguments{
\item{call}{the call whose size shall be determined}
}
\value{
an integer of this size
}
\description{
Find the number of functions and variables (counting multiplies) in a call
}

