% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{count.variables}
\alias{count.variables}
\title{Count all variables appeareances of each variable in a call or expression object}
\usage{
count.variables(call)
}
\arguments{
\item{a}{call object}
}
\value{
table that counts variable names
}
\description{
Count all variables appeareances of each variable in a call or expression object
}

