% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{detibble}
\alias{detibble}
\title{detibble if the input is a tibble it converts it back to a data.frame}
\usage{
detibble(indat)
}
\arguments{
\item{indat}{any matrix like data.frame or tibble}
}
\value{
a data.frame
}
\description{
detibble is used to ensure that if an input object is a tibble
then it is returned as a base R data.frame. This is sometimes needed as
tibbles can upset some base R functionality.
}
\examples{
\dontrun{
# syntax
detibble(tibble)
}
}
