% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract.var.with.index}
\alias{extract.var.with.index}
\title{extracts from a call expression its variable and its index}
\usage{
extract.var.with.index(call, as.character = FALSE)
}
\description{
extracts from a call expression its variable and its index
}

