% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{extractRcode}
\alias{extractRcode}
\title{extractRcode pulls out the r-code blocks from Rmd files}
\usage{
extractRcode(indir, rmdfile, filename = "out.R")
}
\arguments{
\item{indir}{the directory in which the rmd file is to be found and into
which the output file will be placed.}

\item{rmdfile}{name of Rmd/qmd file whose R code is to be extracted}

\item{filename}{the name of the R file into which the r-code is to go.}
}
\value{
generates an R file in the indir but returns nothing
}
\description{
extractRcode pulls out the r-code blocks from Rmd or qmd files
and saves them into a separate R file. This depends upon the prefix \verb{used to define R blocks at the top and bottom. So it 'knows' where they begin and where they end. Any R-code included that is not encased in  the} prefix will not be extracted.
}
\examples{
ddir <- getDBdir()
# syntax   extractRcode(pathtopath(ddir,"projects/AIRF/NWLML"),
#                       "Drivers_of_Productivity.qmd",
#                       filename="productivity.R")
}
