% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{find.funs}
\alias{find.funs}
\title{Find all function calls from a call or expression object}
\usage{
find.funs(call)
}
\value{
unique names of called functions as character vector
}
\description{
Find all function calls from a call or expression object
}

