% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{find.multiple.variables}
\alias{find.multiple.variables}
\title{Find all variables from a call or expression object. 
If a variable appears n times, it is returned n times}
\usage{
find.multiple.variables(call)
}
\arguments{
\item{a}{call object}
}
\value{
variables names (possibly duplicated) as character vector
}
\description{
Find all variables from a call or expression object. 
If a variable appears n times, it is returned n times
}

