% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{find.variables}
\alias{find.variables}
\title{Find all variables from a call or expression object}
\usage{
find.variables(call)
}
\value{
unique variables names as character vector
}
\description{
Find all variables from a call or expression object
}

