% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getConst}
\alias{getConst}
\title{getConst extracts 'nb' numbers from a line of text}
\usage{
getConst(inline, nb, index = 2, split = ",")
}
\arguments{
\item{inline}{text line to be parsed, usually obtained using readLines}

\item{nb}{the number of numbers to extract}

\item{index}{which non-empty object to begin extracting from?}

\item{split}{what split text to use when parsing the line, default = ','}
}
\value{
a vector of length 'nb'
}
\description{
getConst parses a line of text and extracts 'nb' pieces of
text as numbers
}
\examples{
  txtline <- "MaxDL , 32,32,32"
  getConst(txtline,nb=3,index=2)
  txtline <- "bysau, 1, properties by SAU"
  getConst(txtline,nb=1,index=2)
  txtline <- "Dead_Catch_MSY 588.1"
  getConst(txtline,nb=1,index=2,split=" ")
  inline <- "anothername , 31,32,NA"
  getConst(inline=inline,nb=3,index=2)
}
