% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getsingleNum}
\alias{getsingleNum}
\title{getsingleNum find a line of text and extracts a single number}
\usage{
getsingleNum(varname, intxt, split = ",")
}
\arguments{
\item{varname}{the name of the variable to get from intxt}

\item{intxt}{text to be parsed, usually obtained using readLines}

\item{split}{what split text to use when parsing the line, default = ','}
}
\value{
a single number or, if no value is in the data file a NULL
}
\description{
getsingleNum uses grep to find an input line. If the variable
being searched for fails then NULL is returned. It is assumed that the
datafile has been read in using readLines.
}
\examples{
\dontrun{
 txtlines <- c("replicates, 100","Some_other_text, 52")
 getsingleNum("replicates",txtlines)
 getsingleNum("eeplicates",txtlines)
 getsingleNum("other",txtlines)
 txtlines <- "Dead_Catch_MSY 588.1 Another 100"
 getsingleNum("Dead_Catch_MSY",txtlines,split=" ")
}
}
