% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getStr}
\alias{getStr}
\title{getStr obtains a string from an input text line}
\usage{
getStr(inline, sep = ",")
}
\arguments{
\item{inline}{input text line}

\item{sep}{what is used to separate character strings? default = ',', comma}
}
\value{
a vector of character strings
}
\description{
getStr obtains a string from an input text line, omitting the
initial line label. The character to split the line on can be varied
from the default of ','. In the strsplit function the 'fixed' option is
set TRUE so the split will match the sep character exactly.
}
\examples{
txt=c("fleets","trwl","Auln")
getStr(txt)
}
