% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{iscol}
\alias{iscol}
\title{incol determines if a column is present in a matrix or a data.frame.}
\usage{
iscol(incol, inmat)
}
\arguments{
\item{incol}{the name of the column; for example "year"}

\item{inmat}{the matrix or data.frame within which to search for incol}
}
\value{
TRUE or FALSE
}
\description{
incol is a utility to determine whether a named column is
present in a given matrix or data.frame. Is case sensitive and checks
that the inpout is amtrix or data.frame.
}
\examples{
x <- 1:10
test <- matrix(x,nrow=5,ncol=2,dimnames=list(1:5,c("year","Catch")))
print(test)
iscol("year",test)
iscol("Catch",test)
iscol("catch",test)
iscol("ages",test)
}
