% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{limitstr}
\alias{limitstr}
\title{limitstr temporarily constrains the width of the output from str}
\usage{
limitstr(x, width = 72, str1 = TRUE)
}
\arguments{
\item{x}{the object whose structure is to printed}

\item{width}{what width in characters should be used, default= 72}

\item{str1}{should the str1 function be used, which sets max.level=1 and
strips out all the attribute information. default=TRUE}
}
\value{
nothing but it does run str or str1 while leaving the str options
static
}
\description{
limitstr can be used when using Quarto or Rmarkdown to limit
the width of the output text from str or str1 so it fits nicely on an A4
page
}
\examples{
mat <- matrix(data=rnorm(25,mean=20,sd=2),ncol=5,nrow=5,
              dimnames=list(1:5,1:5))
obj <- list(a=mat,b=mat,c=mat)
str(obj)
limitstr(obj,width=40,str1=TRUE)
limitstr(obj,width=40,str1=FALSE)
}
