% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.call}
\alias{make.call}
\title{Creates a call with name name and arguments in arg.list}
\usage{
make.call(name, arg.list, use.names = !is.null(names(arg.list)))
}
\description{
Creates a call with name name and arguments in arg.list
}

