% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{makelist}
\alias{makelist}
\title{makelist a utility that outputs a list structure defined by its input}
\usage{
makelist(scenes)
}
\arguments{
\item{scenes}{an vector of character names describing different scenarios}
}
\value{
a list of length scenes names for the vector of names in scenes
}
\description{
makelist is a utility that performs the common task of making
a list structure of the same length as the vector of names input. It
also names each of the list components after the vector of names. The
output list structure is then ready to be populated with results.
}
\examples{
scenarios <- c("base_case","higher_M","Lower_M")
changeM <- makelist(scenes=scenarios)
changeM
}
