% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pathfinish}
\alias{pathfinish}
\title{pathfinish determines what character is at the end of a path}
\usage{
pathfinish(inpath)
}
\arguments{
\item{inpath}{the path to be analysed}
}
\value{
the end character of the path; either '', '/', or "\\"
}
\description{
pathfinish determines what character is at the end of a
path uses pathkind to get the separator and then checks the end
character
}
\examples{
  indir <- "C:/Users/Malcolm/Dropbox/rcode2/aMSE/data-raw"
  pathfinish(indir)
}
