% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pathfromcopy}
\alias{pathfromcopy}
\title{pathfromcopy converts a path copied in Windows into an R path}
\usage{
pathfromcopy()
}
\value{
a path converted to have forward slashes rather than paired
or single backslashs.
}
\description{
pathfromcopy it is very common to want to convert a path copied
from the file explorer in Windows into an R path. This function uses the
utils functions ,Platform$OS.type and readClipboard tomake this an
automatic process. Its about time I did this!
}
\examples{
# copy any path from the File explorer in Windows, then use:
# pathfromcopy()    # it will read the clipboard and do the conversion.
}
