% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pathkind}
\alias{pathkind}
\title{pathkind finds the type of separator used in a path}
\usage{
pathkind(inpath)
}
\arguments{
\item{inpath}{\itemize{
\item the path to be analysed
}}
}
\value{
the type of path divider, either a 0 = '\\' or a
1 = '/'
}
\description{
pathkind finds the type of separator used in a path,
this is either a '/' or a '\\'
}
\examples{
indir <- "C:/Users/Malcolm/Dropbox/rcode2/aMSE/data-raw"
pathkind(indir)
}
