% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pathstart}
\alias{pathstart}
\title{pathstart determines what character(s) is at the start of a path}
\usage{
pathstart(inpath)
}
\arguments{
\item{inpath}{the path to be analysed}
}
\value{
the start character of the path; either '', '/', or "\\"
}
\description{
pathstart determines what character(s) is at the start of a
path uses pathkind to get the separator and then checks the start
character
}
\examples{
  indir <- "C:/Users/Malcolm/Dropbox/rcode2/aMSE/data-raw"
  pathstart(indir)
}
