% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pathtopath}
\alias{pathtopath}
\title{pathtopath combines two paths accounting for type of separator}
\usage{
pathtopath(path1, path2)
}
\arguments{
\item{path1}{the first path to be added together}

\item{path2}{the second path which is to be added to the first. This may
include a filename is so desired.}
}
\value{
the combined path
}
\description{
pathtopath combines two paths, the second of which could
include a filename, while taking into account the type of separator.
It also prevents a doubling up or missing out of said separator
between the two paths being joined. If the separator type differs
between the two input paths the functions stops with a warning.
}
\examples{
  in1 <- "c:/users/Malcolm/DropBox"
  in2 <- "aMSEUse/scenarios/EG"
  pathtopath(in1, in2)  # no separator character between the paths
  in2 <- "/aMSEUse/scenarios/EG"
  pathtopath(in1, in2)  # path2 with a separator
  in1 <- "c:/users/Malcolm/DropBox/"
  pathtopath(in1, in2)  # both paths with a separator
  in1 <- "c:\\\\users\\\\Malcolm\\\\DropBox"
  in2 <- "aMSEUse\\\\scenarios\\\\EG"
  pathtopath(in1, in2)  # a different separator but missing from both
}
