% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{recursively.replace}
\alias{recursively.replace}
\title{Recursively replace elements of a call or list}
\usage{
recursively.replace(call, replace.fun)
}
\description{
Recursively replace elements of a call or list
}

