% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{strip.parentheses}
\alias{strip.parentheses}
\title{strip a call object from outer parentheses}
\usage{
strip.parentheses(call, parentheses = "(")
}
\description{
strip a call object from outer parentheses
}

