% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codeUtils.R
\name{substitute.call}
\alias{substitute.call}
\title{substitutes in a call object x (works like substitute2 in pryr)}
\usage{
substitute.call(x, env)
}
\description{
substitutes in a call object x (works like substitute2 in pryr)
}

