% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{writepackagedesc}
\alias{writepackagedesc}
\title{writepackagedesc adds a packages R file to the R directory}
\usage{
writepackagedesc(packname = "", packdesc = "")
}
\arguments{
\item{packname}{a character vector of the package's name, default = ''}

\item{packdesc}{a brief description of what the package does, default = ''}
}
\value{
nothing but it does write a function into the R directory, creating
the directory if it does not already exist.
}
\description{
writepackagedesc is only used when creating a new package from
GitHub, as described in ALSimGuide. It must be used when the working
directory (as in getwd()) is the base directory of the incipient
package. Either this functions two arguments are filled or there needs
to be a basic readme.md file created by GitHub from teh package name and
the brief description given inside GitHub when creating the repository.
}
\examples{
# syntax = writepackagedesc(packname="ALSim",packdesc="")
}
