% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find.variables}
\alias{find.variables}
\title{Find all variables from a call or expression object}
\usage{
find.variables(call)
}
\value{
unique variables names as character vector
}
\description{
Find all variables from a call or expression object
}

