% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subst.var}
\alias{subst.var}
\title{Substitute a variable or a symbol in the expression by subs}
\usage{
subst.var(call, var, subs)
}
\arguments{
\item{call}{a call object or string}

\item{var}{a symbol or string}

\item{subs}{a call or string}
}
\description{
Substitute a variable or a symbol in the expression by subs
}

