% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{toXL}
\alias{toXL}
\title{toXL a notice deprecating the function}
\usage{
toXL(x)
}
\arguments{
\item{x}{a dummy variable}
}
\value{
a message descripbing where toXL can now be found
}
\description{
toXL was causing issues for MacOS machines, so I have put the
original function into the source file 'systemcall_source.R' in the
data-raw subdirectory of the package.
}
\examples{
toXL(NULL)
}
